requirements.txt
requests==2.32.3
tenacity==8.2.3
python-dotenv==1.0.1
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
certifi==2024.7.4
